+++ src/main/java/br/eti/emersondantas/api/rebel/services/GetItemsAverageServiceImpl.java
@@ -14,11 +14,11 @@ public class GetItemsAverageServiceImpl implements GetItemsAverageService{
     private final ItemRepository itemRepository;
 
     @Override
-    public HashMap<String, Long> getItemsAverage(){
+    public HashMap<String, Double> getItemsAverage(){
         List<String> names = this.itemRepository.getAllDistinctName();
-        HashMap<String, Long> itemsAmountAvgMap = new HashMap<>();
+        HashMap<String, Double> itemsAmountAvgMap = new HashMap<>();
         for(String itemName: names){
-            Long avg = this.itemRepository.getAverageOfItemsWithThatName(itemName);
+            Double avg = this.itemRepository.getAverageOfItemsWithThatName(itemName);
             itemsAmountAvgMap.put(itemName, avg);
         }
         return itemsAmountAvgMap;
